kurubot: fix conversion issue with map left oob

+++ Misc/KuruBot/KuruBot/Map.cs
@@ -113,8 +113,8 @@ namespace KuruBot
         public bool IsPixelInCollision(int x, int y)
         {
             // Note: these calculus simulate those of the actual Kururin implementation
-            uint xm = (uint)x % (uint)WidthPx;
-            uint ym = (uint)y % (uint)HeightPx;
+            uint xm = (ushort)x % (uint)WidthPx;
+            uint ym = (ushort)y % (uint)HeightPx;
             uint addr = xm + ym * (uint)WidthPx;
             return physical_map[(int)addr];
         }
@@ -123,8 +123,8 @@ namespace KuruBot
             // Note: these calculus simulate those of the actual Kururin implementation
             // C# integer division semantics matches the C one
             // C# modulo semantics matches the C one
-            int x_tile = x / tile_size;
-            int y_tile = y / tile_size;
+            int x_tile = (short)x / tile_size;
+            int y_tile = (short)y / tile_size;
             int xm = x_tile % Width;
             int ym = y_tile % Height;
             ushort tile = TileAt(xm, ym);
